requirements.txt
Django==1.11.7
django-crispy-forms==1.7.0
django-registration-redux==1.9
pytz==2017.3
PyYAML==3.12
ua-parser==0.7.3
user-agents==1.1.0
